refactor: 설문 테스트 패키지 수정
- 설문과 회원 설문 구분을 위해 테스트 패키지 수정
